[idna] Move unit tests to tests/unit.rs
